+++ index.js
@@ -8,6 +8,7 @@
  * Module dependencies.
  */
 
+var accepts = require('accepts')
 var debug = require('debug')('finalhandler')
 var escapeHtml = require('escape-html')
 var http = require('http')
@@ -47,6 +48,8 @@ function finalhandler(req, res, options) {
   var onerror = options.onerror
 
   return function (err) {
+    var body
+    var constructBody
     var msg
 
     // unhandled error
@@ -62,15 +65,12 @@ function finalhandler(req, res, options) {
       }
 
       // production gets a basic error message
-      var msg = env === 'production'
+      msg = env === 'production'
         ? http.STATUS_CODES[res.statusCode]
         : err.stack || err.toString()
-      msg = escapeHtml(msg)
-        .replace(/\n/g, '<br>')
-        .replace(/  /g, ' &nbsp;') + '\n'
     } else {
       res.statusCode = 404
-      msg = 'Cannot ' + escapeHtml(req.method) + ' ' + escapeHtml(req.originalUrl || req.url) + '\n'
+      msg = 'Cannot ' + req.method + ' ' + (req.originalUrl || req.url)
     }
 
     debug('default %s', res.statusCode)
@@ -85,8 +85,76 @@ function finalhandler(req, res, options) {
       return req.socket.destroy()
     }
 
-    send(req, res, res.statusCode, msg)
+    // negotiate
+    var accept = accepts(req)
+    var type = accept.types('html', 'text')
+
+    // construct body
+    switch (type) {
+      case 'html':
+        constructBody = constructHtmlBody
+        break
+      default:
+        // default to plain text
+        constructBody = constructTextBody
+        break
+    }
+
+    // construct body
+    body = constructBody(res.statusCode, msg)
+
+    // send response
+    send(req, res, res.statusCode, body)
   }
+}
+
+/**
+ * Get HTML body string
+ *
+ * @param {number} status
+ * @param {string} message
+ * @return {Buffer}
+ * @api private
+ */
+
+function constructHtmlBody(status, message) {
+  var msg = escapeHtml(message)
+    .replace(/\n/g, '<br>')
+    .replace(/  /g, ' &nbsp;')
+
+  var html = '<!doctype html>\n'
+    + '<html lang=en>\n'
+    + '<head>\n'
+    + '<meta charset=utf-8>\n'
+    + '<title>' + escapeHtml(http.STATUS_CODES[status]) + '</title>\n'
+    + '</head>\n'
+    + '<body>\n'
+    + msg + '\n'
+    + '</body>\n'
+
+  var body = new Buffer(html, 'utf8')
+
+  body.type = 'text/html; charset=utf-8'
+
+  return body
+}
+
+/**
+ * Get plain text body string
+ *
+ * @param {number} status
+ * @param {string} message
+ * @return {Buffer}
+ * @api private
+ */
+
+function constructTextBody(status, message) {
+  var msg = message + '\n'
+  var body = new Buffer(msg, 'utf8')
+
+  body.type = 'text/plain; charset=utf-8'
+
+  return body
 }
 
 /**
@@ -95,7 +163,7 @@ function finalhandler(req, res, options) {
  * @param {IncomingMessage} req
  * @param {OutgoingMessage} res
  * @param {number} status
- * @param {string} body
+ * @param {Buffer} body
  * @api private
  */
 
@@ -107,8 +175,8 @@ function send(req, res, status, body) {
     res.setHeader('X-Content-Type-Options', 'nosniff')
 
     // standard headers
-    res.setHeader('Content-Type', 'text/html; charset=utf-8')
-    res.setHeader('Content-Length', Buffer.byteLength(body, 'utf8'))
+    res.setHeader('Content-Type', body.type)
+    res.setHeader('Content-Length', body.length)
 
     if (req.method === 'HEAD') {
       res.end()