+++ test/test.js
@@ -33,11 +33,11 @@ describe('finalhandler(req, res)', function () {
   })
 
   describe('404 response', function () {
-    it('include method and path', function (done) {
+    it('should include method and path', function (done) {
       var server = createServer()
       request(server)
       .get('/foo')
-      .expect(404, 'Cannot GET /foo\n', done)
+      .expect(404, /Cannot GET \/foo/, done)
     })
 
     it('should handle HEAD', function (done) {
@@ -65,6 +65,28 @@ describe('finalhandler(req, res)', function () {
       test.write(buf)
       test.expect(404, done)
     })
+
+    describe('when HTML acceptable', function () {
+      it('should respond with HTML', function (done) {
+        var server = createServer()
+        request(server)
+        .get('/foo')
+        .set('Accept', 'text/html')
+        .expect('Content-Type', 'text/html; charset=utf-8')
+        .expect(404, /<html/, done)
+      })
+    })
+
+    describe('when HTML not acceptable', function () {
+      it('should respond with plain text', function (done) {
+        var server = createServer()
+        request(server)
+        .get('/foo')
+        .set('Accept', 'application/x-bogus')
+        .expect('Content-Type', 'text/plain; charset=utf-8')
+        .expect(404, 'Cannot GET /foo\n', done)
+      })
+    })
   })
 
   describe('error response', function () {
@@ -72,14 +94,14 @@ describe('finalhandler(req, res)', function () {
       var server = createServer(new Error('boom!'))
       request(server)
       .get('/foo')
-      .expect(500, /^Error: boom!<br> &nbsp; &nbsp;at/, done)
+      .expect(500, /Error: boom!.*at.*:[0-9]+:[0-9]+/, done)
     })
 
     it('should handle HEAD', function (done) {
-      var server = createServer()
+      var server = createServer(new Error('boom!'))
       request(server)
       .head('/foo')
-      .expect(404, '', done)
+      .expect(500, '', done)
     })
 
     it('should include security header', function (done) {
@@ -94,7 +116,7 @@ describe('finalhandler(req, res)', function () {
       var server = createServer('lame string')
       request(server)
       .get('/foo')
-      .expect(500, 'lame string\n', done)
+      .expect(500, /lame string/, done)
     })
 
     it('should send staus code name when production', function (done) {
@@ -103,7 +125,11 @@ describe('finalhandler(req, res)', function () {
       var server = createServer(err, {env: 'production'})
       request(server)
       .get('/foo')
-      .expect(501, 'Not Implemented\n', done)
+      .expect(501, /Not Implemented/, function (err, res) {
+        if (err) return done(err)
+        should(res.text).not.match(/boom!/)
+        done()
+      })
     })
 
     describe('when there is a request body', function () {
@@ -150,6 +176,36 @@ describe('finalhandler(req, res)', function () {
         test.write(buf)
         test.write(buf)
         test.expect(500, done)
+      })
+    })
+
+    describe('when HTML acceptable', function () {
+      it('should respond with HTML', function (done) {
+        var server = createServer(new Error('boom!'))
+        request(server)
+        .get('/foo')
+        .set('Accept', 'text/html')
+        .expect('Content-Type', 'text/html; charset=utf-8')
+        .expect(500, /<html/, done)
+      })
+
+      it('should escape error stack', function (done) {
+        var server = createServer(new Error('boom!'))
+        request(server)
+        .get('/foo')
+        .set('Accept', 'text/html')
+        .expect(500, /Error: boom!<br> &nbsp; &nbsp;at/, done)
+      })
+    })
+
+    describe('when HTML not acceptable', function () {
+      it('should respond with plain text', function (done) {
+        var server = createServer(new Error('boom!'))
+        request(server)
+        .get('/foo')
+        .set('Accept', 'application/x-bogus')
+        .expect('Content-Type', 'text/plain; charset=utf-8')
+        .expect(500, /Error: boom!\n    at/, done)
       })
     })